server/upgrade-tests: Avoid Docker host networking.
Instead, we update the metadata to point to the correct database URL each time we start HGE.

This means we can run the tests on macOS (and any other environment where Docker runs inside a VM).

I was hoping this would solve the issues we're seeing on CI too, but no such luck.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/9528
GitOrigin-RevId: 32cbf65430462b4b0e59f5d644260f7c26c7135c