server: Rename "error-message" to "hasura-error-message".
There is already a library in Hackage called "error-message".

This breaks the Nix build as we end up overriding that library, which causes a change to any of its dependents.

Outside Nix, Shadowing isn't strictly a problem, but it's prudent to avoid it in case we accidentally end up with the other library in edge cases.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/9679
GitOrigin-RevId: e3f54fd4e5eafec9b0a00128767daad0c8aeab72