server: Upgrade ODBC to correct some error reporting.
The ODBC library inserts `CAST(? AS NVARCHAR(MAX))` instead of just `?` when working with text, to ensure that a text parameter is always interpreted as the same type. This was introduced in https://github.com/fpco/odbc/pull/44 to fix a different bug in HGE.

However, this is not reflected in the returned query in case of an error, which can be very confusing for the reader.

This change fixes the renderer so that the query is generated in the same way when parameterized and when rendered for error messages.

The associated changes are at https://github.com/hasura/odbc/commit/43b07813d635215b3466824d07890de452920fc2.

[NDAT-738]: https://hasurahq.atlassian.net/browse/NDAT-738?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/9949
GitOrigin-RevId: 465bab3fccab00d41390b96d4e8ea39f5b140a61