server: Support `_contains` and `_contained_in` in permissions.
We recently added support for the `_contains` and `_contained_in` operators for PostgreSQL arrays. However, they weren't available in permissions.

This adds support for these operators for arrays too.

As these operators are overloaded to work with either JSONB or arrays, no validation is done on the value. If the value is not an array, it will fail when making a query.

[NDAT-806]: https://hasurahq.atlassian.net/browse/NDAT-806?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/10075
GitOrigin-RevId: 02e4a57d39ac3303205c37add4186a587f65e409