server/benchmarks: properly benchmark pro container in EE mode
In f7400814792 we moved to benchmarking the Enterprise Edition pro container, but because we didn't successfully pipe the proper environment variables through to the docker container, we were benchmarking the pro server but in OSS mode. We would rather have visible the performance implications of as much functionality as possible,  so make another attempt to run in EE mode.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/10206
GitOrigin-RevId: 83f823515d82503d1d7f0bf2b19637fad537175b