Alter log level of garbage collector message (fix #10172)
The garbage collector runs periodically to free unused memory with the
default operation set to occur during quiet periods. In the event that
a quiet period does not arise, the garbage collector is forced and a
warning generated to inform users this has occured.

As warnings require users to take action, this log message causes
confusion. Therefore, the log level of the garbage collection message
has been altered from LevelWarn to LevelInfo to provide information
that this has occured without expectation of users to take action.