Deserialize the NDC response once, not twice. (#467)
## Description

This modifies handling of the NDC response so we deserialize it once,
not twice.

The previous code deserialized first to `serde_json::Value`, and then
again to the required type. This is costly and unnecessary. By
parameterizing over the type, we can go directly to the type we're
looking for.

We still want to do the two-step process for errors, but it's probably
fine if they remain a little slower as the error response shouldn't be
too large.

Analysis with flamegraphs before and after shows a dramatic drop in CPU
time as a result.

V3_GIT_ORIGIN_REV_ID: d7574bd8c82f0863372e1463fda020a8174c7341